requirements.txt
Django==3.2.7
jsonschema==3.2.0
django_hosts==4.0
pycryptodome==3.10.1
django-cors-headers==3.8.0
django-bootstrap-v5==1.0.5